+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for OpenEstate-IO
 -   Updated to Java 7.
 -   Updated `commons-lang` library to version 3.6.
 -   Add dependency to `commons-text` library in version 1.1.
+-   Support **OpenImmo** format in version 1.2.7b.
 
 
 1.2 (14 May 2017)